backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis[hiredis]==5.2.1
httpx==0.28.1
anthropic==0.43.0
qdrant-client==1.12.1
stripe==11.4.1
pyyaml==6.0.2
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
python-multipart==0.0.20
grpcio==1.69.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
aiosqlite==0.20.0
fakeredis[json]==2.26.2
